requirements.txt
streamlit>=1.28.0
openai>=1.40.0
pypdf>=4.0.0
python-dotenv>=1.0.0
